+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+
+## 4.17.1
+
+_2023-04-13_
+
 - fix: `getKennitalaBirthday` should return `undefined` for "Invalid Date"s
 
 ## 4.17.0